spring boot upgrade

+++ src/main/java/by/homesite/joplinforwarder/controllers/dto/response/JwtResponse.java
@@ -11,13 +11,13 @@ public class JwtResponse
 {
 	private String accessToken;
 	private String type = "Bearer";
-	private Long id;
+	private Integer id;
 	private String username;
 	private String email;
 	private String lang;
 	private List<String> roles;
 
-	public JwtResponse(String accessToken, Long id, String username, String email, String lang, List<String> roles) {
+	public JwtResponse(String accessToken, Integer id, String username, String email, String lang, List<String> roles) {
 		this.accessToken = accessToken;
 		this.id = id;
 		this.username = username;
